+++ shared/ui/styles/main.less
@@ -177,6 +177,11 @@ body.preload * {
 	input[type="number"],
 	input[type="password"],
 	textarea {
+		padding: 5px 6px;
+		color: @text-color !important;
+		background-color: @app-background-color;
+		border-radius: 2px;
+
 		&.clearable {
 			padding-right: 31px !important;
 		}
